LayoutTests/imported/w3c/web-platform-tests/navigation-api/precommit-handler/multiple-intercept-expected.txt


PASS precommitHandler + precommitHandler assert_equals: expected false but got true
PASS precommitHandler + (not provided) assert_equals: expected false but got true
PASS precommitHandler + handler assert_equals: expected false but got true
PASS handler + precommitHandler assert_equals: expected false but got true
PASS handler + (not provided)
PASS handler + handler
PASS (not provided) + precommitHandler assert_equals: expected false but got true
PASS (not provided) + handler
